pyproject.toml
[project]
name = "semantic-searching"
version = "0.1.0"
description = "Add your description here"
requires-python = "~=3.13.0"
dependencies = [
    "fastapi[standard]",
    "pydantic",
    "numpy",
    "polars",
    "txtai",
    "gensim",
    "scikit-learn",
    "tqdm",
    "torch",
    "haystack-ai",
    "sentence_transformers",
    "transformers<5"
]

[dependency-groups]
build = [
    "click",
    "kagglehub[polars-datasets]",
    "aiohttp[speedups]",
    "beautifulsoup4",
    "lxml"
]
dev = [
    "ipykernel>=7.2.0",
]
